Crate oak_containers_sdk

Re-exports

pub use attestation::InstanceAttester;

pub use attestation::InstanceEndorser;

pub use crypto::InstanceEncryptionKeyHandle;

pub use oak_session_context::ApplicationHandler;

pub use oak_session_context::OakSessionContext;

pub use orchestrator_client::OrchestratorClient;

pub use orchestrator_client::OrchestratorInterface;

Modules

attestation

crypto

handler

oak_session_context

orchestrator_client

standalone

The Oak Standalone SDK allows easy development iteration and testing for an Oak-enabled trusted binary without requiring the entire Oak containers stack.

tonic