Trait oak_containers_sdk::orchestrator_client::OrchestratorInterface

pub trait OrchestratorInterface {
    // Required methods
    fn get_application_config<'life0, 'async_trait>(
        &'life0 mut self,
    ) -> Pin<Box<dyn Future<Output = Result<Vec<u8>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn notify_app_ready<'life0, 'async_trait>(
        &'life0 mut self,
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_endorsed_evidence<'life0, 'async_trait>(
        &'life0 mut self,
    ) -> Pin<Box<dyn Future<Output = Result<EndorsedEvidence>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
}

Trait defining the interface for interacting with the Orchestrator. This trait can be implemented by the real Orchestrator or mocked for testing purposes.

Required Methods

fn get_application_config<'life0, 'async_trait>( &'life0 mut self, ) -> Pin<Box<dyn Future<Output = Result<Vec<u8>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Retrieves the application configuration from the Orchestrator. This configuration contains settings and parameters specific to the application.

fn notify_app_ready<'life0, 'async_trait>( &'life0 mut self, ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Notifies the Orchestrator that the application is ready to receive requests. This should be called after the application has completed its initialization.

fn get_endorsed_evidence<'life0, 'async_trait>( &'life0 mut self, ) -> Pin<Box<dyn Future<Output = Result<EndorsedEvidence>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Retrieves the endorsed evidence from the Orchestrator. This evidence is used to prove the authenticity and integrity of the application.